--- toyproject/src/main/java/com/laonworks/toyboard/Model/Post.java
@@ -1,5 +1,0 @@
-package com.laonworks.toyboard.model;
-
-public class Post {
-
-}
