--- toyproject/src/main/java/com/laonworks/toyboard/Service/MainService.java
@@ -1,9 +1,0 @@
-package com.laonworks.toyboard.service;
-
-import com.laonworks.toyboard.model.Member;
-
-public interface MainService {
-
-	Member getMember(String member_email);
-
-}
