+++ toyproject/src/main/java/com/laonworks/toyboard/service/MainServiceImpl.java
@@ -1,8 +1,10 @@
 package com.laonworks.toyboard.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 
+import com.laonworks.toyboard.dao.MainDao;
 import com.laonworks.toyboard.dao.MainMapper;
 import com.laonworks.toyboard.model.Member;
 
@@ -10,11 +12,20 @@ import com.laonworks.toyboard.model.Member;
 public class MainServiceImpl implements MainService {
 
 	@Autowired
-	private MainMapper md;
-
+	private MainDao md;
+	
+	@Autowired
+	private MainMapper mm;
+	
+	// 회원 검색
+	public Member getMember(String member_email) {
+		return md.getMember(member_email);
+	}
+	
 	/*[비번 찾기 메일 보내기]*/
 	@Override
 	public Member pwfind(Member mem) throws Exception {
-		return md.pwfind(mem);
+		return mm.pwfind(mem);
 	}
+
 }
